Dev (#125)
* refactor scripts

* dev utils

* npm

* fix otel spans ignoring opts and add utility for handlers

* refactor

* update

* add redoc docs and notes for workaround in swagger ui due to tags

* colors and work around swagger ui merge tags

* update some templates

* discriminator

* notes to fix swagger ui not handling anchors

* notes to work around swagger ui

* fix swagger ui shenanigans

* remove tags after yq

* remove old postgen code

* notes

* remove old gen code

* refactor and notes

* refactor

* pgx dates notes

* pgx time.Time notes

* golangci lint

* notes on xo gen structs for type override

* update and notes

* tracer opts

* xo generates sorted create and update params to allow for direct struct conversion from requests

* bring back code that wasnt dead after all

* config fix and automatic linter fixes

* add golangci-lint to scripts

* nullable false and remove required tag for update params

* filter and map

* nullable

* notes for postgen struct gen when schema name differs

* bring back custom schema names resolution in gen-schema

* fix back update_spec_with_structs

* convert slice of structs

* handler services refactor

* uuid pattern

* update - pending discriminator validation and generation conflict resolution

* notes on fixed discriminator

* correct early return and error as check

* simplify workitem create body parsing

* body reading fix

* typo

+++ internal/repos/postgresql/postgresqltestutil/notification.go
@@ -9,7 +9,7 @@ import (
 	"github.com/google/uuid"
 )
 
-func RandomNotificationCreateParams(t *testing.T, receiverRank *int16, sender uuid.UUID, receiver *uuid.UUID, notificationType db.NotificationType) *db.NotificationCreateParams {
+func RandomNotificationCreateParams(t *testing.T, receiverRank *int, sender uuid.UUID, receiver *uuid.UUID, notificationType db.NotificationType) *db.NotificationCreateParams {
 	t.Helper()
 
 	return &db.NotificationCreateParams{
