+++ internal/repos/postgresql/work_item_comment.go
@@ -31,7 +31,7 @@ func (wit *WorkItemComment) Create(ctx context.Context, d db.DBTX, params *db.Wo
 	return workItemComment, nil
 }
 
-func (wit *WorkItemComment) Update(ctx context.Context, d db.DBTX, id int64, params *db.WorkItemCommentUpdateParams) (*db.WorkItemComment, error) {
+func (wit *WorkItemComment) Update(ctx context.Context, d db.DBTX, id int, params *db.WorkItemCommentUpdateParams) (*db.WorkItemComment, error) {
 	workItemComment, err := wit.ByID(ctx, d, id)
 	if err != nil {
 		return nil, fmt.Errorf("could not get workItemComment by id %w", parseErrorDetail(err))
@@ -47,7 +47,7 @@ func (wit *WorkItemComment) Update(ctx context.Context, d db.DBTX, id int64, par
 	return workItemComment, err
 }
 
-func (wit *WorkItemComment) ByID(ctx context.Context, d db.DBTX, id int64) (*db.WorkItemComment, error) {
+func (wit *WorkItemComment) ByID(ctx context.Context, d db.DBTX, id int) (*db.WorkItemComment, error) {
 	workItemComment, err := db.WorkItemCommentByWorkItemCommentID(ctx, d, id)
 	if err != nil {
 		return nil, fmt.Errorf("could not get workItemComment: %w", parseErrorDetail(err))
@@ -56,7 +56,7 @@ func (wit *WorkItemComment) ByID(ctx context.Context, d db.DBTX, id int64) (*db.
 	return workItemComment, nil
 }
 
-func (wit *WorkItemComment) Delete(ctx context.Context, d db.DBTX, id int64) (*db.WorkItemComment, error) {
+func (wit *WorkItemComment) Delete(ctx context.Context, d db.DBTX, id int) (*db.WorkItemComment, error) {
 	workItemComment, err := wit.ByID(ctx, d, id)
 	if err != nil {
 		return nil, fmt.Errorf("could not get workItemComment by id %w", parseErrorDetail(err))
