+++ internal/repos/reposwrappers/timeentry_with_otel.gen.go
@@ -36,7 +36,7 @@ func NewTimeEntryWithTracing(base repos.TimeEntry, instance string, spanDecorato
 }
 
 // ByID implements repos.TimeEntry
-func (_d TimeEntryWithTracing) ByID(ctx context.Context, d db.DBTX, id int64) (tp1 *db.TimeEntry, err error) {
+func (_d TimeEntryWithTracing) ByID(ctx context.Context, d db.DBTX, id int) (tp1 *db.TimeEntry, err error) {
 	ctx, _span := otel.Tracer(_d._instance).Start(ctx, "repos.TimeEntry.ByID")
 	defer func() {
 		if _d._spanDecorator != nil {
@@ -84,7 +84,7 @@ func (_d TimeEntryWithTracing) Create(ctx context.Context, d db.DBTX, params *db
 }
 
 // Delete implements repos.TimeEntry
-func (_d TimeEntryWithTracing) Delete(ctx context.Context, d db.DBTX, id int64) (tp1 *db.TimeEntry, err error) {
+func (_d TimeEntryWithTracing) Delete(ctx context.Context, d db.DBTX, id int) (tp1 *db.TimeEntry, err error) {
 	ctx, _span := otel.Tracer(_d._instance).Start(ctx, "repos.TimeEntry.Delete")
 	defer func() {
 		if _d._spanDecorator != nil {
@@ -108,7 +108,7 @@ func (_d TimeEntryWithTracing) Delete(ctx context.Context, d db.DBTX, id int64) 
 }
 
 // Update implements repos.TimeEntry
-func (_d TimeEntryWithTracing) Update(ctx context.Context, d db.DBTX, id int64, params *db.TimeEntryUpdateParams) (tp1 *db.TimeEntry, err error) {
+func (_d TimeEntryWithTracing) Update(ctx context.Context, d db.DBTX, id int, params *db.TimeEntryUpdateParams) (tp1 *db.TimeEntry, err error) {
 	ctx, _span := otel.Tracer(_d._instance).Start(ctx, "repos.TimeEntry.Update")
 	defer func() {
 		if _d._spanDecorator != nil {
